refactor: move csv deletion into its own file
Resolves #300
