+++ app/src/main/java/ai/elimu/analytics/task/DeleteOldCsvFilesWorker.kt
@@ -1,0 +1,60 @@
+package ai.elimu.analytics.task
+
+import ai.elimu.analytics.BuildConfig
+import android.content.Context
+import androidx.work.Worker
+import androidx.work.WorkerParameters
+import timber.log.Timber
+import java.io.File
+
+/**
+ * Deletes CSV files that were generated by previous versions of the app.
+ */
+class DeleteOldCsvFilesWorker(context: Context, workerParams: WorkerParameters) :
+    Worker(context, workerParams) {
+
+    companion object {
+        private const val DIR_LETTER_SOUND_LEARNING_EVENTS = "letter-sound-learning-events"
+        private const val DIR_LETTER_SOUND_ASSESSMENT_EVENTS = "letter-sound-assessment-events"
+        private const val DIR_WORD_LEARNING_EVENTS = "word-learning-events"
+        private const val DIR_WORD_ASSESSMENT_EVENTS = "word-assessment-events"
+        private const val DIR_NUMBER_LEARNING_EVENTS = "number-learning-events"
+        private const val DIR_NUMBER_ASSESSMENT_EVENTS = "number-assessment-events"
+        private const val DIR_STORYBOOK_LEARNING_EVENTS = "storybook-learning-events"
+        private const val DIR_VIDEO_LEARNING_EVENTS = "video-learning-events"
+    }
+
+    override fun doWork(): Result {
+        Timber.i("doWork")
+
+        val eventSubDirs = listOf(
+            DIR_LETTER_SOUND_LEARNING_EVENTS,
+            DIR_LETTER_SOUND_ASSESSMENT_EVENTS,
+            DIR_WORD_LEARNING_EVENTS,
+            DIR_WORD_ASSESSMENT_EVENTS,
+            DIR_NUMBER_LEARNING_EVENTS,
+            DIR_NUMBER_ASSESSMENT_EVENTS,
+            DIR_STORYBOOK_LEARNING_EVENTS,
+            DIR_VIDEO_LEARNING_EVENTS
+        )
+        val filesDir = applicationContext.filesDir
+        val languageDirs = filesDir.listFiles { file ->
+            file.isDirectory && file.name.startsWith("lang-")
+        } ?: emptyArray()
+
+        for (languageDir in languageDirs) {
+            for (eventSubDir in eventSubDirs) {
+                val eventsDir = File(languageDir, eventSubDir)
+                for (csvFile in eventsDir.listFiles() ?: emptyArray()) {
+                    if (!csvFile.name.contains("_${BuildConfig.VERSION_CODE}_") && csvFile.name.endsWith(".csv")) {
+                        val deleted = csvFile.delete()
+                        Timber.i("delete ${csvFile.name}: $deleted")
+                    }
+                }
+            }
+        }
+
+        Timber.i("Cleanup complete!")
+        return Result.success()
+    }
+}
