+++ app/src/main/java/ai/elimu/analytics/task/TaskInitializer.kt
@@ -37,5 +37,15 @@ object TaskInitializer {
             ExistingPeriodicWorkPolicy.KEEP,
             uploadEventsWorkRequest
         )
+
+        // Periodically delete CSV files that were generated by previous versions of the app
+        val deleteOldCsvFilesWorkRequest = PeriodicWorkRequest
+            .Builder(DeleteOldCsvFilesWorker::class.java, 1, TimeUnit.DAYS)
+            .build()
+        WorkManager.getInstance(context).enqueueUniquePeriodicWork(
+            "delete_old_csv_files",
+            ExistingPeriodicWorkPolicy.KEEP,
+            deleteOldCsvFilesWorkRequest
+        )
     }
 }
