+++ app/src/main/java/ai/elimu/analytics/task/UploadEventsWorker.kt
@@ -25,16 +25,6 @@ import java.io.File
  */
 class UploadEventsWorker(context: Context, workerParams: WorkerParameters) :
     Worker(context, workerParams) {
-    companion object {
-        private const val DIR_LETTER_SOUND_LEARNING_EVENTS = "letter-sound-learning-events"
-        private const val DIR_LETTER_SOUND_ASSESSMENT_EVENTS = "letter-sound-assessment-events"
-        private const val DIR_WORD_LEARNING_EVENTS = "word-learning-events"
-        private const val DIR_WORD_ASSESSMENT_EVENTS = "word-assessment-events"
-        private const val DIR_NUMBER_LEARNING_EVENTS = "number-learning-events"
-        private const val DIR_NUMBER_ASSESSMENT_EVENTS = "number-assessment-events"
-        private const val DIR_STORYBOOK_LEARNING_EVENTS = "storybook-learning-events"
-        private const val DIR_VIDEO_LEARNING_EVENTS = "video-learning-events"
-    }
 
     override fun doWork(): Result {
         Timber.i("doWork")
@@ -51,43 +41,8 @@ class UploadEventsWorker(context: Context, workerParams: WorkerParameters) :
         uploadStoryBookLearningEvents()
 
         uploadVideoLearningEvents()
-
-        deleteCsvFilesFromPreviousVersions()
         
         return Result.success()
-    }
-
-    private fun deleteCsvFilesFromPreviousVersions() {
-        Timber.i("deleteCsvFilesFromPreviousVersions")
-
-        val eventSubDirs = listOf(
-            DIR_LETTER_SOUND_LEARNING_EVENTS,
-            DIR_LETTER_SOUND_ASSESSMENT_EVENTS,
-            DIR_WORD_LEARNING_EVENTS,
-            DIR_WORD_ASSESSMENT_EVENTS,
-            DIR_NUMBER_LEARNING_EVENTS,
-            DIR_NUMBER_ASSESSMENT_EVENTS,
-            DIR_STORYBOOK_LEARNING_EVENTS,
-            DIR_VIDEO_LEARNING_EVENTS
-        )
-        val filesDir = applicationContext.filesDir
-        val languageDirs = filesDir.listFiles { file ->
-            file.isDirectory && file.name.startsWith("lang-")
-        } ?: emptyArray()
-
-        for (languageDir in languageDirs) {
-            for (eventSubDir in eventSubDirs) {
-                val eventsDir = File(languageDir, eventSubDir)
-                for (csvFile in eventsDir.listFiles() ?: emptyArray()) {
-                    if (!csvFile.name.contains("_${BuildConfig.VERSION_CODE}_") && csvFile.name.endsWith(".csv")) {
-                        val deleted = csvFile.delete()
-                        Timber.i("delete ${csvFile.name}: $deleted")
-                    }
-                }
-            }
-        }
-
-        Timber.i("deleteCsvFilesFromPreviousVersions complete!")
     }
 
     // Upload CSV files generated by the current version of the app, e.g:
@@ -97,7 +52,7 @@ class UploadEventsWorker(context: Context, workerParams: WorkerParameters) :
         Timber.i("uploadLetterSoundLearningEvents")
 
         val languageDir = File(applicationContext.filesDir, "lang-${SharedPreferencesHelper.getLanguage(applicationContext)}")
-        val eventsDir = File(languageDir, DIR_LETTER_SOUND_LEARNING_EVENTS)
+        val eventsDir = File(languageDir, "letter-sound-learning-events")
         Timber.i("eventsDir.exists(): ${eventsDir.exists()}")
         for (csvFile in eventsDir.listFiles() ?: emptyArray()) {
             Timber.i("csvFile: ${csvFile}")
@@ -140,7 +95,7 @@ class UploadEventsWorker(context: Context, workerParams: WorkerParameters) :
         Timber.i("uploadLetterSoundAssessmentEvents")
 
         val languageDir = File(applicationContext.filesDir, "lang-${SharedPreferencesHelper.getLanguage(applicationContext)}")
-        val eventsDir = File(languageDir, DIR_LETTER_SOUND_ASSESSMENT_EVENTS)
+        val eventsDir = File(languageDir, "letter-sound-assessment-events")
         Timber.i("eventsDir.exists(): ${eventsDir.exists()}")
         for (csvFile in eventsDir.listFiles() ?: emptyArray()) {
             Timber.i("csvFile: ${csvFile}")
@@ -184,7 +139,7 @@ class UploadEventsWorker(context: Context, workerParams: WorkerParameters) :
         Timber.i("uploadWordLearningEvents")
 
         val languageDir = File(applicationContext.filesDir, "lang-${SharedPreferencesHelper.getLanguage(applicationContext)}")
-        val eventsDir = File(languageDir, DIR_WORD_LEARNING_EVENTS)
+        val eventsDir = File(languageDir, "word-learning-events")
         Timber.i("eventsDir.exists(): ${eventsDir.exists()}")
         for (csvFile in eventsDir.listFiles() ?: emptyArray()) {
             Timber.i("csvFile: ${csvFile}")
@@ -227,7 +182,7 @@ class UploadEventsWorker(context: Context, workerParams: WorkerParameters) :
         Timber.i("uploadWordAssessmentEvents")
 
         val languageDir = File(applicationContext.filesDir, "lang-${SharedPreferencesHelper.getLanguage(applicationContext)}")
-        val eventsDir = File(languageDir, DIR_WORD_ASSESSMENT_EVENTS)
+        val eventsDir = File(languageDir, "word-assessment-events")
         Timber.i("eventsDir.exists(): ${eventsDir.exists()}")
         for (csvFile in eventsDir.listFiles() ?: emptyArray()) {
             Timber.i("csvFile: ${csvFile}")
@@ -271,7 +226,7 @@ class UploadEventsWorker(context: Context, workerParams: WorkerParameters) :
         Timber.i("uploadNumberLearningEvents")
 
         val languageDir = File(applicationContext.filesDir, "lang-${SharedPreferencesHelper.getLanguage(applicationContext)}")
-        val eventsDir = File(languageDir, DIR_NUMBER_LEARNING_EVENTS)
+        val eventsDir = File(languageDir, "number-learning-events")
         Timber.i("eventsDir.exists(): ${eventsDir.exists()}")
         for (csvFile in eventsDir.listFiles() ?: emptyArray()) {
             Timber.i("csvFile: ${csvFile}")
@@ -314,7 +269,7 @@ class UploadEventsWorker(context: Context, workerParams: WorkerParameters) :
         Timber.i("uploadNumberAssessmentEvents")
 
         val languageDir = File(applicationContext.filesDir, "lang-${SharedPreferencesHelper.getLanguage(applicationContext)}")
-        val eventsDir = File(languageDir, DIR_NUMBER_ASSESSMENT_EVENTS)
+        val eventsDir = File(languageDir, "number-assessment-events")
         Timber.i("eventsDir.exists(): ${eventsDir.exists()}")
         for (csvFile in eventsDir.listFiles() ?: emptyArray()) {
             Timber.i("csvFile: ${csvFile}")
@@ -358,7 +313,7 @@ class UploadEventsWorker(context: Context, workerParams: WorkerParameters) :
         Timber.i("uploadStoryBookLearningEvents")
 
         val languageDir = File(applicationContext.filesDir, "lang-${SharedPreferencesHelper.getLanguage(applicationContext)}")
-        val eventsDir = File(languageDir, DIR_STORYBOOK_LEARNING_EVENTS)
+        val eventsDir = File(languageDir, "storybook-learning-events")
         Timber.i("eventsDir.exists(): ${eventsDir.exists()}")
         for (csvFile in eventsDir.listFiles() ?: emptyArray()) {
             Timber.i("csvFile: ${csvFile}")
@@ -402,7 +357,7 @@ class UploadEventsWorker(context: Context, workerParams: WorkerParameters) :
         Timber.i("uploadVideoLearningEvents")
 
         val languageDir = File(applicationContext.filesDir, "lang-${SharedPreferencesHelper.getLanguage(applicationContext)}")
-        val eventsDir = File(languageDir, DIR_VIDEO_LEARNING_EVENTS)
+        val eventsDir = File(languageDir, "video-learning-events")
         Timber.i("eventsDir.exists(): ${eventsDir.exists()}")
         for (csvFile in eventsDir.listFiles() ?: emptyArray()) {
             Timber.i("csvFile: ${csvFile}")
